pyproject.toml
[project]
name = "ai-benchmark"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "gputil>=1.4.0",
    "matplotlib>=3.10.0",
    "numpy>=2.0.2",
    "pandas>=2.2.3",
    "python-dotenv>=1.0.1",
    "scikit-learn>=1.6.1",
    "tensorflow>=2.18.0",
    "tensorflow-metal>=1.2.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
